Include guards: fix comment skipping
The attempt to make comment skipping more robust caused a regression, because
`beginning-of-defun' also goes to the start of a comment. Use the old
`forward-sexp' + `backward-sexp' heuristic instead, even though it fails if a
header starts with a delimited statement (e.g. in\\nt). This is pretty obscure
though.

Finding #endif should not suffer from that problem.

Move the code to go the end of the buffer skipping comments into a separate
function, even though it is used only once, so that such functions can be close
together (`emr-cc--beggining-of-header' and `emr-cc--end-of-header').